+++ exp/apps/compiler_outputs/pagerank_pull-topological_edge-cut/gen_cuda.cuh
@@ -14,13 +14,13 @@
 struct CUDA_Context {
 	int device;
 	int id;
-	size_t nowned;
+	unsigned int nowned;
 	CSRGraphTy hg;
 	CSRGraphTy gg;
-	size_t *num_master_nodes; // per host
-	Shared<size_t> *master_nodes; // per host
-	size_t *num_slave_nodes; // per host
-	Shared<size_t> *slave_nodes; // per host
+	unsigned int *num_master_nodes; // per host
+	Shared<unsigned int> *master_nodes; // per host
+	unsigned int *num_slave_nodes; // per host
+	Shared<unsigned int> *slave_nodes; // per host
 	Shared<int> nout;
 	Shared<int> *master_nout; // per host
 	Shared<int> *slave_nout; // per host
@@ -52,7 +52,7 @@ void min_node_nout_cuda(struct CUDA_Context *ctx, unsigned LID, int v) {
 		nout[LID] = v;
 }
 
-__global__ void batch_get_node_nout(index_type size, size_t * p_master_nodes, int * p_master_nout, int * p_nout) {
+__global__ void batch_get_node_nout(index_type size, unsigned int * p_master_nodes, int * p_master_nout, int * p_nout) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -71,7 +71,7 @@ void batch_get_node_nout_cuda(struct CUDA_Context *ctx, unsigned from_id, int *v
 	memcpy(v, ctx->master_nout[from_id].cpu_rd_ptr(), sizeof(int) * ctx->num_master_nodes[from_id]);
 }
 
-__global__ void batch_get_reset_node_nout(index_type size, size_t * p_slave_nodes, int * p_slave_nout, int * p_nout, int value) {
+__global__ void batch_get_reset_node_nout(index_type size, unsigned int * p_slave_nodes, int * p_slave_nout, int * p_nout, int value) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -91,7 +91,7 @@ void batch_get_reset_node_nout_cuda(struct CUDA_Context *ctx, unsigned from_id, 
 	memcpy(v, ctx->slave_nout[from_id].cpu_rd_ptr(), sizeof(int) * ctx->num_slave_nodes[from_id]);
 }
 
-__global__ void batch_set_node_nout(index_type size, size_t * p_slave_nodes, int * p_slave_nout, int * p_nout) {
+__global__ void batch_set_node_nout(index_type size, unsigned int * p_slave_nodes, int * p_slave_nout, int * p_nout) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -110,7 +110,7 @@ void batch_set_node_nout_cuda(struct CUDA_Context *ctx, unsigned from_id, int *v
 	check_cuda_kernel;
 }
 
-__global__ void batch_add_node_nout(index_type size, size_t * p_master_nodes, int * p_master_nout, int * p_nout) {
+__global__ void batch_add_node_nout(index_type size, unsigned int * p_master_nodes, int * p_master_nout, int * p_nout) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -129,7 +129,7 @@ void batch_add_node_nout_cuda(struct CUDA_Context *ctx, unsigned from_id, int *v
 	check_cuda_kernel;
 }
 
-__global__ void batch_min_node_nout(index_type size, size_t * p_master_nodes, int * p_master_nout, int * p_nout) {
+__global__ void batch_min_node_nout(index_type size, unsigned int * p_master_nodes, int * p_master_nout, int * p_nout) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -169,7 +169,7 @@ void min_node_value_cuda(struct CUDA_Context *ctx, unsigned LID, float v) {
 		value[LID] = v;
 }
 
-__global__ void batch_get_node_value(index_type size, size_t * p_master_nodes, float * p_master_value, float * p_value) {
+__global__ void batch_get_node_value(index_type size, unsigned int * p_master_nodes, float * p_master_value, float * p_value) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -188,7 +188,7 @@ void batch_get_node_value_cuda(struct CUDA_Context *ctx, unsigned from_id, float
 	memcpy(v, ctx->master_value[from_id].cpu_rd_ptr(), sizeof(float) * ctx->num_master_nodes[from_id]);
 }
 
-__global__ void batch_get_reset_node_value(index_type size, size_t * p_slave_nodes, float * p_slave_value, float * p_value, float value) {
+__global__ void batch_get_reset_node_value(index_type size, unsigned int * p_slave_nodes, float * p_slave_value, float * p_value, float value) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -208,7 +208,7 @@ void batch_get_reset_node_value_cuda(struct CUDA_Context *ctx, unsigned from_id,
 	memcpy(v, ctx->slave_value[from_id].cpu_rd_ptr(), sizeof(float) * ctx->num_slave_nodes[from_id]);
 }
 
-__global__ void batch_set_node_value(index_type size, size_t * p_slave_nodes, float * p_slave_value, float * p_value) {
+__global__ void batch_set_node_value(index_type size, unsigned int * p_slave_nodes, float * p_slave_value, float * p_value) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -227,7 +227,7 @@ void batch_set_node_value_cuda(struct CUDA_Context *ctx, unsigned from_id, float
 	check_cuda_kernel;
 }
 
-__global__ void batch_add_node_value(index_type size, size_t * p_master_nodes, float * p_master_value, float * p_value) {
+__global__ void batch_add_node_value(index_type size, unsigned int * p_master_nodes, float * p_master_value, float * p_value) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -246,7 +246,7 @@ void batch_add_node_value_cuda(struct CUDA_Context *ctx, unsigned from_id, float
 	check_cuda_kernel;
 }
 
-__global__ void batch_min_node_value(index_type size, size_t * p_master_nodes, float * p_master_value, float * p_value) {
+__global__ void batch_min_node_value(index_type size, unsigned int * p_master_nodes, float * p_master_value, float * p_value) {
 	unsigned tid = TID_1D;
 	unsigned nthreads = TOTAL_THREADS_1D;
 	index_type src_end = size;
@@ -305,28 +305,28 @@ void load_graph_CUDA(struct CUDA_Context *ctx, MarshalGraph &g, unsigned num_hos
 	memcpy(graph.edge_dst, g.edge_dst, sizeof(index_type) * g.nedges);
 	if(g.node_data) memcpy(graph.node_data, g.node_data, sizeof(node_data_type) * g.nnodes);
 	if(g.edge_data) memcpy(graph.edge_data, g.edge_data, sizeof(edge_data_type) * g.nedges);
-	ctx->num_master_nodes = (size_t *) calloc(num_hosts, sizeof(size_t));
-	memcpy(ctx->num_master_nodes, g.num_master_nodes, sizeof(size_t) * num_hosts);
-	ctx->master_nodes = (Shared<size_t> *) calloc(num_hosts, sizeof(Shared<size_t>));
+	ctx->num_master_nodes = (unsigned int *) calloc(num_hosts, sizeof(unsigned int));
+	memcpy(ctx->num_master_nodes, g.num_master_nodes, sizeof(unsigned int) * num_hosts);
+	ctx->master_nodes = (Shared<unsigned int> *) calloc(num_hosts, sizeof(Shared<unsigned int>));
 	ctx->master_nout = (Shared<int> *) calloc(num_hosts, sizeof(Shared<int>));
 	ctx->master_value = (Shared<float> *) calloc(num_hosts, sizeof(Shared<float>));
 	for(uint32_t h = 0; h < num_hosts; ++h){
 		if (ctx->num_master_nodes[h] > 0) {
 			ctx->master_nodes[h].alloc(ctx->num_master_nodes[h]);
-			memcpy(ctx->master_nodes[h].cpu_wr_ptr(), g.master_nodes[h], sizeof(size_t) * ctx->num_master_nodes[h]);
+			memcpy(ctx->master_nodes[h].cpu_wr_ptr(), g.master_nodes[h], sizeof(unsigned int) * ctx->num_master_nodes[h]);
 			ctx->master_nout[h].alloc(ctx->num_master_nodes[h]);
 			ctx->master_value[h].alloc(ctx->num_master_nodes[h]);
 		}
 	}
-	ctx->num_slave_nodes = (size_t *) calloc(num_hosts, sizeof(size_t));
-	memcpy(ctx->num_slave_nodes, g.num_slave_nodes, sizeof(size_t) * num_hosts);
-	ctx->slave_nodes = (Shared<size_t> *) calloc(num_hosts, sizeof(Shared<size_t>));
+	ctx->num_slave_nodes = (unsigned int *) calloc(num_hosts, sizeof(unsigned int));
+	memcpy(ctx->num_slave_nodes, g.num_slave_nodes, sizeof(unsigned int) * num_hosts);
+	ctx->slave_nodes = (Shared<unsigned int> *) calloc(num_hosts, sizeof(Shared<unsigned int>));
 	ctx->slave_nout = (Shared<int> *) calloc(num_hosts, sizeof(Shared<int>));
 	ctx->slave_value = (Shared<float> *) calloc(num_hosts, sizeof(Shared<float>));
 	for(uint32_t h = 0; h < num_hosts; ++h){
 		if (ctx->num_slave_nodes[h] > 0) {
 			ctx->slave_nodes[h].alloc(ctx->num_slave_nodes[h]);
-			memcpy(ctx->slave_nodes[h].cpu_wr_ptr(), g.slave_nodes[h], sizeof(size_t) * ctx->num_slave_nodes[h]);
+			memcpy(ctx->slave_nodes[h].cpu_wr_ptr(), g.slave_nodes[h], sizeof(unsigned int) * ctx->num_slave_nodes[h]);
 			ctx->slave_nout[h].alloc(ctx->num_slave_nodes[h]);
 			ctx->slave_value[h].alloc(ctx->num_slave_nodes[h]);
 		}
