+++ libexp/include/Galois/Runtime/Cuda/cuda_mtypes.h
@@ -13,15 +13,15 @@ typedef unsigned int edge_data_type;
 struct MarshalGraph {
   size_t nnodes;  
   size_t nedges;
-  size_t nowned;
+  unsigned int nowned;
   int id;
   index_type *row_start;
   index_type *edge_dst;
   node_data_type *node_data;
   edge_data_type *edge_data;
-  size_t *num_master_nodes;
-  size_t **master_nodes;
-  size_t *num_slave_nodes;
-  size_t **slave_nodes;
+  unsigned int *num_master_nodes;
+  unsigned int **master_nodes;
+  unsigned int *num_slave_nodes;
+  unsigned int **slave_nodes;
 };
